+++ lib/domain/repositories/event_impact_repository.dart
@@ -1,0 +1,5 @@
+import '../entities/event_impact.dart';
+
+abstract interface class EventImpactRepository {
+  Future<List<EventImpact>> loadEventImpacts({required String parkId});
+}